rename mixins and events
