+++ src/main/java/darkorg/betterdurability/mixin/vanilla/LivingEntityMixin.java
@@ -1,4 +1,4 @@
-package darkorg.betterdurability.mixin;
+package darkorg.betterdurability.mixin.vanilla;
 
 import darkorg.betterdurability.util.ItemUtil;
 import darkorg.betterdurability.util.StackUtil;
